### 2022-09-24 #4792 @editorialbot
[PRE REVIEW]: rigr: Regression, Inference, and General Data Analysis Tools in R
**Submitting author:** @yiqunchen (<a href="http://orcid.org/0000-0002-4100-1507">Yiqun Chen</a>)
**Repository:** https://github.com/statdivlab/rigr
**Branch with paper.md** (empty if default branch): joss_submission
**Version:** v1.0.4
