### 2022-10-07 #4792 @osorensen
👋 @n8thangreen @RBhupi @stefanocoretta, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
