### 2022-10-13 #4847 @editorialbot
[REVIEW]: rigr: Regression, Inference, and General Data Analysis Tools in R
**Submitting author:** @yiqunchen (<a href="http://orcid.org/0000-0002-4100-1507">Yiqun Chen</a>)
**Repository:** https://github.com/statdivlab/rigr
**Branch with paper.md** (empty if default branch): joss_submission
**Version:** v.1.0.5
