### 2022-10-30 #4847 @tomsing1
Great R package, @yiqunchen et al - and congratulations to getting it released on CRAN!
@osorensen : looks great to me, I didn't spot any issues. My review is complete.
