+++ README.md
@@ -1,11 +1,8 @@
-
 <!-- README.md is generated from README.Rmd. Please edit that file -->
+
 <!-- badges: start  -->
 
-[![CRAN
-status](https://www.r-pkg.org/badges/version/rigr)](https://CRAN.R-project.org/package=rigr)
-[![R-CMD-check](https://github.com/statdivlab/rigr/workflows/R-CMD-check/badge.svg)](https://github.com/statdivlab/rigr/actions)
-[![codecov](https://codecov.io/gh/statdivlab/rigr/branch/main/graph/badge.svg)](https://app.codecov.io/gh/statdivlab/rigr)
+[![CRAN status](https://www.r-pkg.org/badges/version/rigr)](https://CRAN.R-project.org/package=rigr) [![R-CMD-check](https://github.com/statdivlab/rigr/workflows/R-CMD-check/badge.svg)](https://github.com/statdivlab/rigr/actions) [![codecov](https://codecov.io/gh/statdivlab/rigr/branch/main/graph/badge.svg)](https://app.codecov.io/gh/statdivlab/rigr)
 
 <!-- badges: end -->
 
@@ -13,22 +10,9 @@ status](https://www.r-pkg.org/badges/version/rigr)](https://CRAN.R-project.org/p
 
 ## Introduction
 
-`rigr` is an `R` package to streamline data analysis in `R`. Learning
-both `R` and introductory statistics at the same time can be
-challenging, and so we created `rigr` to facilitate common data analysis
-tasks and enable learners to focus on statistical concepts.
+`rigr` is an `R` package to streamline data analysis in `R`. Learning both `R` and introductory statistics at the same time can be challenging, and so we created `rigr` to facilitate common data analysis tasks and enable learners to focus on statistical concepts.
 
-`rigr`, formerly known as
-[`uwIntroStats`](https://CRAN.R-project.org/package=uwIntroStats),
-provides easy-to-use interfaces for descriptive statistics, one- and
-two-sample inference, and regression analyses. `rigr` output includes
-key information while omitting unnecessary details that can be confusing
-to beginners. Heteroskedasticity-robust (“sandwich”) standard errors are
-returned by default, and multiple partial F-tests and tests for
-contrasts are easy to specify. A single regression function
-(`regress()`) can fit both linear models, generalized linear models, and
-proportional hazards models, allowing students to more easily make
-connections between different classes of models.
+`rigr`, formerly known as [`uwIntroStats`](https://CRAN.R-project.org/package=uwIntroStats), provides easy-to-use interfaces for descriptive statistics, one- and two-sample inference, and regression analyses. `rigr` output includes key information while omitting unnecessary details that can be confusing to beginners. Heteroskedasticity-robust (“sandwich”) standard errors are returned by default, and multiple partial F-tests and tests for contrasts are easy to specify. A single regression function (`regress()`) can fit both linear models, generalized linear models, and proportional hazards models, allowing students to more easily make connections between different classes of models.
 
 ## Installation
 
@@ -36,49 +20,33 @@ You can install the stable release of `rigr` from CRAN as follows:
 
     install.packages("rigr")
 
-You can install the development version of `rigr` from GitHub using the
-code below.
+You can install the development version of `rigr` from GitHub using the code below. The installment is through the `R` package `remotes`.
 
+    # run the following code to make sure that the remote package it installed
+    if (!requireNamespace("remotes", quietly = TRUE)) {
+      install.package("remotes")
+    }
+    # install the package
     remotes::install_github("statdivlab/rigr")
 
-If this produces an error, please run `install.packages("remotes")`
[…]
