### 2022-12-07 statdivlab/rigr#138 @osorensen
Editorial comments on JOSS submission
I'm reading through the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4847), and will post comments as issues in the repository.

My first comment is what you write on lines 66-67: "In the following example, we use regress to fit a linear regression model of atrophy (a measure of global brain activity assessed using MRI) on sex, weight, age, and race."

The part in the parentheses does not seem correct, since atrophy in fact is a measure of tissue loss.

There are a few more things I don't understand about the example application starting on line 62:

- Is the atrophy measured by volume, area, or thickness?
- Is it cortical atrophy, averaged over the cerebral cortex, or does it also include subcortical volumes?
- Atrophy is a measure of change, but estimating change requires at least two timepoints. What is then actually represented by the variable `atrophy` in the `mri` dataset?

I looked this up in the Kuller et al. (2007), and I wonder if what is actually represented by the `atrophy` variable is white matter grade? Could you please clarify, and if necessary update both the paper and the `mri` example dataset?
