### 2022-12-07 statdivlab/rigr#139 @osorensen
Editorial comments on JOSS submission
Here are some further comments on the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4847).

- [ ] The reference to R specifies the year 2013. Could you please update so the year matches the version of R used to create the examples shown in the paper? The `citation()` command in `R` should return what you need.
- [ ] Please add DOIs to all paper references where a DOI exists. This at least includes the ones highlighted below:

![image](https://user-images.githubusercontent.com/21175639/206158729-10134389-9c09-4732-97e1-eaef47807f23.png)
