+++ paper.bib
@@ -6,7 +6,8 @@
   number={9},
   pages={1307--1315},
   year={2007},
-  publisher={Elsevier}
+  publisher={Elsevier},
+    doi = {10.1016/j.neurobiolaging.2006.06.010}
 }
 
 @Book{rothman2008modern,
@@ -27,7 +28,9 @@
   number    =  8486,
   pages     = "908--909",
   month     =  apr,
-  year      =  1986
+  year      =  1986,
+    doi = {10.1016/S0140-6736(86)92832-1}
+
 }
 
  @Manual{stats-package,
@@ -72,22 +75,21 @@
   }
 
 @ARTICLE{Balding2006-fq,
-  title     = "A tutorial on statistical methods for population association
-               studies",
+  title     = "A tutorial on statistical methods for population association studies",
   author    = "Balding, David J",
-  journal   = "Nature reviews. Genetics",
+  journal   = "Nature Reviews Genetics",
   publisher = "Nature Publishing Group",
   volume    =  7,
   number    =  10,
   pages     = "781--791",
   month     =  oct,
   year      =  2006,
-  language  = "en"
+  language  = "en",
+  doi = {10.1038/nrg1916}
 }
 
 @ARTICLE{Kim2018-xv,
-  title     = "Joint Analysis of Multiple Interaction Parameters in Genetic
-               Association Studies",
+  title     = "Joint Analysis of Multiple Interaction Parameters in Genetic Association Studies",
   author    = "Kim, Jihye and Ziyatdinov, Andrey and Laville, Vincent and Hu,
                Frank B and Rimm, Eric and Kraft, Peter and Aschard, Hugues",
   journal   = "Genetics",
@@ -97,7 +99,8 @@
   pages     = "483--494",
   month     =  dec,
   year      =  2018,
-  language  = "en"
+  language  = "en",
+  doi = {10.1534/genetics.118.301394}
 }
 
 
@@ -109,18 +112,18 @@
   number={2},
   pages={159--179},
   year={2015},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  doi = {10.1093/pan/mpu015}
 }
 @ARTICLE{Mansournia2021-cd,
-  title    = "Reflection on modern methods: demystifying robust standard errors
-              for epidemiologists",
-  author   = "Mansournia, Mohammad Ali and Nazemipour, Maryam and Naimi, Ashley
-              I and Collins, Gary S and Campbell, Michael J",
-  journal  = "International journal of epidemiology",
+  title    = "Reflection on modern methods: demystifying robust standard errors for epidemiologists",
+  author   = "Mansournia, Mohammad Ali and Nazemipour, Maryam and Naimi, Ashley I and Collins, Gary S and Campbell, Michael J",
+  journal  = "International Journal of Epidemiology",
   volume   =  50,
   number   =  1,
   pages    = "346--351",
   month    =  mar,
   year     =  2021,
-  language = "en"
+  language = "en",
+  doi = {10.1093/ije/dyaa260}
 }
