### 2022-12-17 statdivlab/rigr#138 @yiqunchen
Thanks for your detailed read and clarification! The MRI dataset was originally derived from Kuller et al. (2007) but has been adapted for the use of this package.

In this case, atrophy is measured by the degree of ventricular enlargement relative to the predicted ventricular size; with 0 indicating no ventricular enlargement and 100 indicating the most severe degree of atrophy.

We have clarified the text in the submission as follows:

In the following example, we use `regress` to fit a linear regression model of atrophy (a measure between 0 and 100 of global brain activity measured by the degree of ventricular enlargement relative to the predicted ventricular size; with 0 indicating no ventricular enlargement and 100 indicating the most severe degree of atrophy) on sex, weight, age, and race.

Please let us know if you have any other questions; thank you again for your thorough review!
