+++ paper.md
@@ -65,7 +65,7 @@ We designed the `rigr` package to lower the barriers to obtaining robust and int
 
 We first demonstrate the use of the `regress` function, which is a single function that can be used to fit linear, logistic, Poisson, and proportional hazards regressions depending on the specified `fnctl` argument.
 
-In the following example, we use `regress` to fit a linear regression model of atrophy (a measure of global brain activity assessed using MRI) on sex, weight, age, and race. We see that the robust standard error estimates are reported by default, as are the results of an *F*-test for the null hypothesis that there is no difference in mean atrophy across populations matched on sex, weight and age but who differ in their self-reported racial categories. The `mri` dataset represents a subset of the Cardiovascular Health Study dataset [@kuller2007white]. In order to preserve patient anonymity and facilitate its use in the classroom, the data were modified by adding random noise and rescaling some variables. Such modifications preserved the general relationships among the variables in the dataset.
+In the following example, we use `regress` to fit a linear regression model of atrophy (a measure between 0 and 100 of global brain activity measured by the degree of ventricular enlargement relative to the predicted ventricular size; with 0 indicating no ventricular enlargement and 100 indicating the most severe degree of atrophy) on sex, weight, age, and race. We see that the robust standard error estimates are reported by default, as are the results of an *F*-test for the null hypothesis that there is no difference in mean atrophy messure across populations matched on sex, weight and age but who differ in their self-reported racial categories. The `mri` dataset represents a subset of the Cardiovascular Health Study dataset [@kuller2007white;@mri_dataset]. In order to preserve patient anonymity and facilitate its use in the classroom, the data were modified by adding random noise and rescaling some variables. Such modifications preserved the general relationships among the variables in the dataset.
 
 ```r
 # Loading library and dataset
