+++ paper.bib
@@ -10,6 +10,14 @@
     doi = {10.1016/j.neurobiolaging.2006.06.010}
 }
 
+
+@Manual{mri_dataset,
+    title = {Documentation for {MRI} and {Cerebral Atrophy}},
+   author = {Scott S. Emerson},
+    year = {2005},
+    url = {http://www.emersonstatistics.com/datasets/mri.pdf},
+ }
+
 @Book{rothman2008modern,
   title={{M}odern {E}pidemiology},
   author={Rothman, Kenneth J and Greenland, Sander and Lash, Timothy L},
@@ -33,16 +41,16 @@
 
 }
 
+
+
  @Manual{stats-package,
     title = {R: A Language and Environment for Statistical Computing},
     author = {{R Core Team}},
     organization = {R Foundation for Statistical Computing},
     address = {Vienna, Austria},
-    year = {2013},
-    note = {{ISBN} 3-900051-07-0},
-    url = {http://www.R-project.org/},
+    year = {2021},
+    url = {https://www.R-project.org/},
   }
-
 
  @Article{sandwich-package,
     title = {Various Versatile Variances: An Object-Oriented
