### 2022-12-17 statdivlab/rigr#139 @yiqunchen
This has been updated and fixed in https://github.com/statdivlab/rigr/commit/4ba0416fa87affe0f8e8842387a1133ad230d996 -- thank you for your prompt and detailed review!!
