### 2022-12-17 statdivlab/rigr#138 @osorensen
Thanks for addressing this issue!

I'm still not completely on board, however, since the text still suggests that atrophy is a measure of global brain activity, which it is not.

I suggest rewriting to the following:
> In the following example, we use regress to fit a linear regression model of atrophy (a measure between 0 and 100 of **loss of neurons estimated**  by the degree of ventricular enlargement relative to the predicted ventricular size; with 0 indicating no ventricular enlargement and 100 indicating the most severe degree of atrophy) on sex, weight, age, and race.

See, e.g., [here](https://www.ninds.nih.gov/health-information/disorders/cerebral-atrophy) or [here](https://www.britannica.com/science/atrophy).
