### 2022-12-18 cea238c
fix atrophy descrip
