### 2022-12-18 statdivlab/rigr#138 @yiqunchen
Thank you for your great suggestion! In this revision ([3ccb5a2](https://github.com/statdivlab/rigr/commit/3ccb5a24b50a119ae9bed03ad8c010d30a52c347)), we have edited the sentence to:

In the following example, we use `regress` to fit a linear regression model of atrophy (a measure of loss of neurons estimated by the degree of ventricular enlargement relative to the predicted ventricular size; with 0 indicating no atrophy and 100 indicating the most severe degree of atrophy) on sex, weight, age, and race.

We really appreciate your reviews!
