### 2022-12-18 statdivlab/rigr#138 @osorensen
Great! Could you please also update the documentation of the `atrophy` variable in `mri` dataset accordingly?

That is, here: https://github.com/statdivlab/rigr/blob/0d5f6eb2fa82bd80b3e41fadcdcb90a6694102eb/R/data.R#L95
