+++ .github/workflows/code-coverage.yml
@@ -1,6 +1,6 @@
 on:
   push:
-    branches: [main, master, summer2022]
+    branches: [main, master]
   pull_request:
     branches: [main, master]
 
@@ -15,7 +15,7 @@ jobs:
     steps:
       - uses: actions/checkout@v2
 
-      - uses: r-lib/actions/setup-r@v1
+      - uses: r-lib/actions/setup-r@v2
         with:
           use-public-rspm: true
 
