### 2022-12-19 statdivlab/rigr#138 @yiqunchen
Thank you for your suggestion -- it has been revised accordingly in https://github.com/statdivlab/rigr/blob/main/R/data.R#:~:text=%27%20%5Citem%7Batrophy%7D%7Ba,degree%20of%20atrophy.%7D
