+++ pkgdown.yml
@@ -1,11 +1,11 @@
 pandoc: 2.19.2
-pkgdown: 2.0.6
+pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   descrip_intro: descrip_intro.html
   one_and_two_sample_inference: one_and_two_sample_inference.html
   regress_intro: regress_intro.html
-last_built: 2022-11-07T17:06Z
+last_built: 2022-12-19T07:20Z
 urls:
   reference: https://statdivlab.github.io/rigr/reference
   article: https://statdivlab.github.io/rigr/articles
