+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   descrip_intro: descrip_intro.html
   one_and_two_sample_inference: one_and_two_sample_inference.html
   regress_intro: regress_intro.html
-last_built: 2022-12-19T07:23Z
+last_built: 2022-12-19T08:20Z
 urls:
   reference: https://statdivlab.github.io/rigr/reference
   article: https://statdivlab.github.io/rigr/articles
