### 2022-12-19 #4847 @yiqunchen
Thank you @osorensen for your prompt reviews and actions throughout the review process!! Please see below for the response:
 * [x]  Make a tagged release of your software, and list the version tag of the archived version here: v.1.0.5 (https://github.com/statdivlab/rigr/releases/tag/v.1.0.5)
* [x]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository): https://zenodo.org/record/7456326#.Y6Ac7OzMI6A
* [x]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID: On behalf of the authors, I confirm that the metadata is correct and people who are listed as a co-author of this paper has made significant contributions to the software.
 * [x]  Please list the DOI of the archived version here:  [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7456326.svg)](https://doi.org/10.5281/zenodo.7456326)

Thank you again for your consideration of our submission to JOSS!
