### 2022-12-19 #4847 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
