### 2023-01-03 #4847 @yiqunchen
Happy new year and thank you everyone for your prompt, detailed, and helpful feedback!! It's been an amazing submission process at JOSS!
